pyproject.toml
[tool.poetry]
name = "bdrpca"
version = "0.1.0"
description = ""
authors = ["Nathan Graule <solarliner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
scipy = "^1.9.3"
Pillow = "^9.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
